@@ -693,9 +693,9 @@ namespace PlaywrightSharp.Tests.Page
             }");
 
             var exception = await Assert.ThrowsAsync<TimeoutException>(()
-                => Page.ClickAsync("button", timeout: 5000));
+                => Page.ClickAsync("button", timeout: 3000));
 
-            Assert.Contains("Timeout 5000ms exceeded", exception.Message);
+            Assert.Contains("Timeout 3000ms exceeded", exception.Message);
             Assert.Contains("waiting for element to be visible, enabled and not moving", exception.Message);
             Assert.Contains("element is moving - waiting", exception.Message);
         }
