@@ -41,7 +41,7 @@ namespace PlaywrightSharp
         BackgroundSync,
 
         /// <summary>
-        /// Ambient light sensor, Accelerometer, Gyroscope, Magnetometer
+        /// Ambient light sensor, Accelerometer, Gyroscope, Magnetometer.
         /// </summary>
         Sensors,
 
