@@ -8,19 +8,19 @@ namespace PlaywrightSharp
     public enum LifecycleEvent
     {
         /// <summary>
-        /// Consider navigation to be finished when the <c>load</c> event is fired
+        /// Consider navigation to be finished when the <c>load</c> event is fired.
         /// </summary>
         [EnumMember(Value = "load")]
         Load,
 
         /// <summary>
-        /// Consider navigation to be finished when the <c>DOMContentLoaded</c> event is fired
+        /// Consider navigation to be finished when the <c>DOMContentLoaded</c> event is fired.
         /// </summary>
         [EnumMember(Value = "domcontentloaded")]
         DOMContentLoaded,
 
         /// <summary>
-        /// Consider navigation to be finished when there are no more than 0 network connections for at least <c>500</c> ms
+        /// Consider navigation to be finished when there are no more than 0 network connections for at least <c>500</c> ms.
         /// </summary>
         [EnumMember(Value = "networkidle")]
         Networkidle,
